requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
openpyxl==3.1.2
python-decouple==3.8
django-cors-headers==4.3.0
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
python-jose==3.3.0
cryptography==41.0.7
xlrd==2.0.1
gunicorn==21.2.0
whitenoise==6.6.0